search_api/requirements.txt
# Server
fastapi
uvicorn[standard]
gunicorn
structlog

# ML & Data
pyserini
numpy
psycopg2-binary
pgvector

# General
python-dotenv

# Authentication
python-jose[cryptography]
python-multipart
httpx
